Revert "Merge branch 'main' of https://github.com/dathd6/GECCO2019---Bi-objective-Traveling-Thief-Competition"
This reverts commit ff691e43dd504e90ebc2bfbe379a1dc98deb86d0, reversing
changes made to 5d350757d2167052130f5a4bebfb080304909678.

+++ !EVALUATION/non_dominated_sorting.py
@@ -16,7 +16,7 @@ def fast_non_dominated_sort(F, **kwargs):
 
     # final rank that will be returned
     n_ranked = 0
-    ranked = np.zeros(n)
+    ranked = np.zeros(n, dtype=np.int)
 
     # for each individual a list of all individuals that are dominated by this one
     is_dominating = [[] for _ in range(n)]
